refactor(types): move store types to type folder and update es6 imports
