--- src/store/actions/index.ts
@@ -1,2 +1,0 @@
-export * from './application'
-export * from './reviews'
