+++ src/store/actions/reviews.ts
@@ -1,5 +1,41 @@
-import type { ActionType, BasicAction } from '@app/store'
-import type { Review, ReviewFormValues } from '@app/types'
+import type {
+  ActionType,
+  BasicAction,
+  Review,
+  ReviewFormValues
+} from '@app/types'
+
+export interface AddReviewAction extends BasicAction {
+  type: 'ADD_REVIEW',
+  payload: {
+    reviewForm: ReviewFormValues
+  }
+}
+
+export const ADD_REVIEW: ActionType = 'ADD_REVIEW'
+
+interface AddReview {
+  (reviewForm: ReviewFormValues): AddReviewAction
+}
+export const onAddReview: AddReview = reviewForm => ({
+  type: 'ADD_REVIEW',
+  payload: {
+    reviewForm
+  }
+})
+
+export interface GetReviewAction extends BasicAction {
+  type: 'GET_REVIEW'
+}
+
+export const GET_REVIEW: ActionType = 'GET_REVIEW'
+
+interface GetReview {
+  (): GetReviewAction
+}
+export const onGetReview: GetReview = () => ({
+  type: 'GET_REVIEW'
+})
 
 export interface ReviewFetchStartedAction extends BasicAction {
   type: 'REVIEW_FETCH_STARTED'
@@ -49,38 +85,5 @@ export const onReviewFetched: ReviewFetched = review => ({
   type: 'REVIEW_FETCHED',
   payload: {
     review
-  }
-})
-
-// /////////////////
-export interface GetReviewAction extends BasicAction {
-  type: 'GET_REVIEW'
-}
-
-export const GET_REVIEW: ActionType = 'GET_REVIEW'
-
-interface GetReview {
-  (): GetReviewAction
-}
-export const onGetReview: GetReview = () => ({
-  type: 'GET_REVIEW'
-})
-
-export interface AddReviewAction extends BasicAction {
-  type: 'ADD_REVIEW',
-  payload: {
-    reviewForm: ReviewFormValues
-  }
-}
-
-export const ADD_REVIEW: ActionType = 'ADD_REVIEW'
-
-interface AddReview {
-  (reviewForm: ReviewFormValues): AddReviewAction
-}
-export const onAddReview: AddReview = reviewForm => ({
-  type: 'ADD_REVIEW',
-  payload: {
-    reviewForm
   }
 })
