+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from './device'
 export * from './reviewForm'
 export * from './reviews'
 export * from './routes'
+export * from './store'
